fix gid overflow

+++ DangerousD/GameCore/Managers/MapManager.cs
@@ -45,7 +45,7 @@ namespace DangerousD.GameCore.Managers;
         }
 
         
-        private TileSet GetTileSet(int gid)
+        private TileSet GetTileSet(long gid)
         {
             if (_tileSets == null || _tileSets.Count == 0)
             {
@@ -95,7 +95,7 @@ namespace DangerousD.GameCore.Managers;
                 string[] tiles = chunk.InnerText.Split(',');
                 for (int i = 0; i < tiles.Length; i++)
                 {
-                    int gid = int.Parse(tiles[i]);
+                    long gid = long.Parse(tiles[i]);
 
                     if (gid == 0) continue;
                     
@@ -105,7 +105,7 @@ namespace DangerousD.GameCore.Managers;
                     pos *= _scale;
                     
                     
-                    Rectangle sourceRect = new(new Point((gid - tileSet.FirstGid) % tileSet.Columns, (gid - tileSet.FirstGid) / tileSet.Columns) * tileSet.TileSize, tileSet.TileSize);
+                    Rectangle sourceRect = new(new Point((int)(gid - tileSet.FirstGid) % tileSet.Columns, (int)(gid - tileSet.FirstGid) / tileSet.Columns) * tileSet.TileSize, tileSet.TileSize);
                     Type type = Type.GetType($"DangerousD.GameCore.GameObjects.{tileType}");
                     Activator.CreateInstance(type, pos, tileSet.TileSize.ToVector2() * _scale, sourceRect);
                 }
@@ -131,7 +131,7 @@ namespace DangerousD.GameCore.Managers;
                 Entity inst;
                 if (typeof(Door).IsAssignableFrom(type))
                 {
-                    int gid = entity.Attributes["gid"] is not null ? int.Parse(entity.Attributes["gid"].Value) : 0;
+                    long gid = entity.Attributes["gid"] is not null ? int.Parse(entity.Attributes["gid"].Value) : 0;
                     TileSet tileSet = GetTileSet(gid);
                     Vector2 objectSize = new(int.Parse(entity.Attributes["width"].Value), int.Parse(entity.Attributes["height"].Value));
                     
@@ -143,7 +143,7 @@ namespace DangerousD.GameCore.Managers;
                         if (level is not null)  // Teleport to level
                         {
                             inst = new TeleportingDoor(pos, objectSize,
-                                new Rectangle(new Point((gid - tileSet.FirstGid) * tileSet.TileSize.X, 0),
+                                new Rectangle(new Point((int)(gid - tileSet.FirstGid) * tileSet.TileSize.X, 0),
                                     tileSet.TileSize), () => {AppManager.Instance.ChangeMap(level.Attributes["value"].Value, GetStartCoordinates(level.Attributes["value"].Value));});
                         }
                         else  // Teleport to location
@@ -153,13 +153,13 @@ namespace DangerousD.GameCore.Managers;
                             XmlNode dest = layer.SelectSingleNode($"object[@id = '{target}']");
                         
                             
-                            inst = new TeleportingDoor(pos, objectSize, new Rectangle(new Point((gid - tileSet.FirstGid) * tileSet.TileSize.X, 0), tileSet.TileSize),
+                            inst = new TeleportingDoor(pos, objectSize, new Rectangle(new Point((int)(gid - tileSet.FirstGid) * tileSet.TileSize.X, 0), tileSet.TileSize),
                                 new Vector2(float.Parse(dest.Attributes["x"].Value) + offsetX,
                                     float.Parse(dest.Attributes["y"].Value) + offsetY) * _scale);
                         }
                     } 
                     else
-                        inst = new Door(pos, objectSize, new Rectangle(new Point((gid - tileSet.FirstGid) * tileSet.TileSize.X, 0), tileSet.TileSize));
+                        inst = new Door(pos, objectSize, new Rectangle(new Point((int)(gid - tileSet.FirstGid) * tileSet.TileSize.X, 0), tileSet.TileSize));
                 }
                 else
                 {
